pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hdr_forge"
version = "0.4.0"
description = "HDR Forge video converter - Convert HDR10, Dolby Vision, and SDR videos to H.265/HEVC"
readme = "README.md"
requires-python = ">=3.7"
license = {text = "MIT"}
authors = [
    {name = "Jan Kremser", email = "git@nerdbrief.de"}
]
keywords = ["video", "hdr", "dolby-vision", "h265", "hevc", "converter", "ffmpeg"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Video :: Conversion",
]

[project.urls]
Homepage = "https://github.com/jankremser/hdr_forge"
Repository = "https://github.com/jankremser/hdr_forge"
Issues = "https://github.com/jankremser/hdr_forge/issues"

[project.optional-dependencies]
gui = ["Pillow>=9.0.0"]

[project.scripts]
hdr_forge = "hdr_forge.main:main"
hdr_forge_ui = "hdr_forge.main_ui:main_ui"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
